[RUM-22934][RUM-23164][BUG] Paint timings not supported in Firefox
