+++ src/core/performance/NavigationV2PerformanceService.js
@@ -1,28 +1,7 @@
-import PerformanceService from './PerformanceService';
 import NormalizationHelper from '../Normalization.helper';
+import NavigationV1PerformanceService from './NavigationV1PerformanceService';
 
-export default class PerformanceNavigationPaintTimingService extends PerformanceService {
-  /**
-   * Return the status of the Performance Service
-   * @returns {string}
-   */
-  getStatus() {
-    const timings = [
-      !!this.getResponseStart(),
-      !!this.getDOMInteractive(),
-      !!this.getDOMInteractive(),
-      !!this.getDOMComplete(),
-      !!this.getLoadEventEnd(),
-      !!this.getFirstPaint(),
-      !!this.getFirstContentfulPaint()
-    ];
-
-    if (!timings.filter((i) => { return i === false; }).length) {
-      return this.STATUS.COMPLETE;
-    }
-
-    return this.STATUS.PENDING;
-  }
+export default class NavigationV2PerformanceService extends NavigationV1PerformanceService {
 
   /**
    * Return performance navigation data
@@ -30,16 +9,6 @@ export default class PerformanceNavigationPaintTimingService extends Performance
    */
   getNavigationData() {
     return window.performance.getEntriesByType('navigation')[0];
-  }
-
-  /**
-   * Return performance paint data
-   * @returns {object}
-   */
-  getPaintData(type) {
-    return window.performance.getEntriesByType('paint').filter((i) => {
-      return i.name === type;
-    })[0];
   }
 
   /**
@@ -106,34 +75,6 @@ export default class PerformanceNavigationPaintTimingService extends Performance
     try {
       return NormalizationHelper.normalizeNonZeroPositiveInteger(
         this.getNavigationData().loadEventEnd
-      );
-    } catch (e) {
-      return null;
-    }
-  }
-
-  /**
-   * Return timing duration
-   * @returns {number|null}
-   */
-  getFirstPaint() {
-    try {
-      return NormalizationHelper.normalizeNonZeroPositiveInteger(
-        this.getPaintData('first-paint').startTime
-      );
-    } catch (e) {
-      return null;
-    }
-  }
-
-  /**
-   * Return timing duration
-   * @returns {number|null}
-   */
-  getFirstContentfulPaint() {
-    try {
-      return NormalizationHelper.normalizeNonZeroPositiveInteger(
-        this.getPaintData('first-contentful-paint').startTime
       );
     } catch (e) {
       return null;
